requirements.txt
# FastAPI framework and related dependencies
fastapi
uvicorn
requests
python-dotenv

# OpenAI API for NLP chatbot functionalities
openai

matplotlib


